Added Author template
